+++ test/core/QCheck2_tests.ml
@@ -181,6 +181,14 @@ module Generator = struct
     Test.make ~name:"char has right range'" ~count:1000 ~print:Print.char
       Gen.char (fun c -> '\000' <= c && c <= '\255')
 
+  let printable_test =
+    Test.make ~name:"printable has right range" ~count:1000 ~print:Print.char
+      Gen.printable (fun c -> c = '\n' || 32 <= Char.code c && Char.code c <= 126)
+
+  let numeral_test =
+    Test.make ~name:"numeral has right range" ~count:1000 ~print:Print.char
+      Gen.numeral (fun c -> '0' <= c && c <= '9')
+
   let nat_test =
     Test.make ~name:"nat has right range" ~count:1000 ~print:Print.int
       Gen.nat (fun n -> 0 <= n && n < 10000)
@@ -369,6 +377,14 @@ module Shrink = struct
   let char_is_never_abcdef =
     Test.make ~name:"char never produces 'abcdef'" ~count:1000 ~print:Print.char
       Gen.char (fun c -> not (List.mem c ['a';'b';'c';'d';'e';'f']))
+
+  let printable_is_never_sign = (* should shrink towards '!' with lowest ascii code 33 *)
+    Test.make ~name:"printable never produces '!\"#$%&''" ~count:1000 ~print:Print.char
+      Gen.printable (fun c -> not (List.mem c ['!';'@';'#';'$';'%']))
+
+  let numeral_is_never_less_5 =
+    Test.make ~name:"printable never produces less than '5" ~count:1000 ~print:Print.char
+      Gen.numeral (fun c -> c >= '5')
 
   let strings_are_empty =
     Test.make ~name:"strings are empty" ~count:1000 ~print:Print.string
@@ -581,6 +597,8 @@ module Shrink = struct
     ints_smaller_209609;
     nats_smaller_5001;
     char_is_never_abcdef;
+    printable_is_never_sign;
+    numeral_is_never_less_5;
     strings_are_empty;
     string_never_has_000_char;
     string_never_has_255_char;
@@ -745,8 +763,12 @@ module Stats = struct
   let bool_dist =
     Test.make ~name:"bool dist" ~count:500_000 ~collect:Bool.to_string Gen.bool (fun _ -> true)
 
-  let char_dist =
-    Test.make ~name:"char code dist" ~count:500_000 ~stats:[("char code", Char.code)] Gen.char (fun _ -> true)
+  let char_dist_tests =
+    [
+    Test.make ~name:"char code dist"           ~count:500_000 ~stats:[("char code", Char.code)] Gen.char      (fun _ -> true);
+    Test.make ~name:"printable char code dist" ~count:500_000 ~stats:[("char code", Char.code)] Gen.printable (fun _ -> true);
+    Test.make ~name:"numeral char code dist"   ~count:500_000 ~stats:[("char code", Char.code)] Gen.numeral   (fun _ -> true);
+  ]
 
   let string_len_tests =
     let len = ("len",String.length) in
@@ -819,9 +841,9 @@ module Stats = struct
       Gen.(oneof [small_int_corners ();int]) (fun _ -> true)
 
   let tests =
-    [ bool_dist;
-      char_dist;
-      tree_depth_test;]
+    [ bool_dist; ]
+    @ char_dist_tests
+    @ [ tree_depth_test;]
     @ string_len_tests
     @ [pair_dist;
        triple_dist;